requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
openai
beautifulsoup4
python-docx
openpyxl
PyPDF2
python-dotenv
